[refactor] modularize core (ruby and js) for better maintenance
